scripts/requirements.txt
supabase>=2.0.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0
